Fix literal equality
Before, a literal (say, `"a"`) would match any value *beginning* with
that string (say, `"a"`, `"ab"`, and `"abc"`). Now, it correctly only
matches `"a"`.

Before, in the `gte`, we ended up comparing the next byte in the key
with the `boundary`, which is a high value. Thus, looking for `"a"`,
we'd find `"ab"`, because `"b" < boundary`. Instead, this case needs the
`separator + boundary` used in other cases, because `"b" > separator`.